# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Escrituras",
        "description": "API para acceder a las escrituras de La Iglesia de Jesucristo de los Santos de los Últimos Días",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/capitulos/{id}": {
            "get": {
                "tags": [
                    "Capitulos"
                ],
                "summary": "Obtiene un capítulo específico",
                "operationId": "95148e269e3580ad88b333105e1c219e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del capítulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Capítulo encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Capitulo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Capítulo no encontrado"
                    }
                }
            }
        },
        "/api/capitulos/{id}/pericopas": {
            "get": {
                "tags": [
                    "Capitulos"
                ],
                "summary": "Lista todas las pericopas de un capítulo",
                "operationId": "973043407c8208bc88556bdcb516b911",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del capítulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pericopas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pericopa"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Capítulo no encontrado"
                    }
                }
            }
        },
        "/api/capitulos/{id}/versiculos": {
            "get": {
                "tags": [
                    "Capitulos"
                ],
                "summary": "Lista todos los versículos de un capítulo",
                "operationId": "e5a6162a3c2d86dab8fcec2e4d75a1cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del capítulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de versículos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Versiculo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Capítulo no encontrado"
                    }
                }
            }
        },
        "/api/capitulos/{id}/parte": {
            "get": {
                "tags": [
                    "Capitulos"
                ],
                "summary": "Obtiene la parte a la que pertenece el capítulo",
                "operationId": "18f11d12206e32e83f33b2aff9aed0dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del capítulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parte obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Parte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Capítulo no encontrado"
                    }
                }
            }
        },
        "/api/capitulos/{id}/libro": {
            "get": {
                "tags": [
                    "Capitulos"
                ],
                "summary": "Obtiene el libro al que pertenece el capítulo",
                "operationId": "ebe7663581a330d0aff88387f81e77c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del capítulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Libro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Capítulo no encontrado"
                    }
                }
            }
        },
...
        "/api/divisiones/{id}": {
            "get": {
                "tags": [
                    "Divisiones"
                ],
                "summary": "Obtiene una división específica",
                "operationId": "c3cf42cc74eb52f8c73bae4043c88ac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la división",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "División encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Division"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "División no encontrada"
                    }
                }
            }
        },
        "/api/divisiones/{id}/libros": {
            "get": {
                "tags": [
                    "Divisiones"
                ],
                "summary": "Lista todos los libros de una división",
                "operationId": "cc6df11012446b4fa397fdd12a8f1e06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la división",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de libros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Libro"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "División no encontrada"
                    }
                }
            }
        },
        "/api/divisiones/{id}/volumen": {
            "get": {
                "tags": [
                    "Divisiones"
                ],
                "summary": "Obtiene el volumen al que pertenece la división",
                "operationId": "12b7c770c148c3f03ad79a8d5928c815",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la división",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Volumen obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Volumen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "División no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Volumen": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Division": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "volumen_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Libro": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "division_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Parte": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "libro_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Capitulo": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "num_capitulo": {
                        "type": "integer"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "libro_id": {
                        "type": "integer"
                    },
                    "parte_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Pericopa": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "titulo": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "capitulo_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Versiculo": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "contenido": {
                        "type": "string"
                    },
                    "num_versiculo": {
                        "type": "integer"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "capitulo_id": {
                        "type": "integer"
                    },
                    "pericopa_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}